remove: 不要なファイルの削除とアセットのパスを更新
