+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "expo-app",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "src/screens/app.tsx",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
