+++ src/assets/icon.png
