+++ src/screens/app.tsx
@@ -1,0 +1,23 @@
+import { registerRootComponent } from "expo";
+import { StatusBar } from "expo-status-bar";
+import { StyleSheet, Text, View } from "react-native";
+
+const App = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.tsx to start working on your app!!!!!!!</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
+
+registerRootComponent(App);
